+++ src/main/java/com/dwarfeng/ftp/struct/FtpConfig.java
@@ -1,5 +1,8 @@
 package com.dwarfeng.ftp.struct;
 
+import com.dwarfeng.dutil.basic.prog.Buildable;
+import com.dwarfeng.ftp.util.FtpConfigUtil;
+
 /**
  * FTP 配置。
  *
@@ -21,6 +24,17 @@ public final class FtpConfig {
             String host, int port, String username, String password, String serverCharset, int connectTimeout,
             long noopInterval, int bufferSize
     ) {
+        // 检查参数。
+        FtpConfigUtil.checkHost(host);
+        FtpConfigUtil.checkPort(port);
+        FtpConfigUtil.checkUsername(username);
+        FtpConfigUtil.checkPassword(password);
+        FtpConfigUtil.checkServerCharset(serverCharset);
+        FtpConfigUtil.checkConnectTimeout(connectTimeout);
+        FtpConfigUtil.checkNoopInterval(noopInterval, connectTimeout);
+        FtpConfigUtil.checkBufferSize(bufferSize);
+
+        // 设置值。
         this.host = host;
         this.port = port;
         this.username = username;
@@ -75,5 +89,79 @@ public final class FtpConfig {
                 ", noopInterval=" + noopInterval +
                 ", bufferSize=" + bufferSize +
                 '}';
+    }
+
+    /**
+     * FTP 配置构造器。
+     *
+     * @author DwArFeng
+     * @since 1.1.9
+     */
+    public static final class Builder implements Buildable<FtpConfig> {
+
+        public static final int DEFAULT_PORT = 21;
+        public static final String DEFAULT_SERVER_CHARSET = "UTF-8";
+        public static final int DEFAULT_CONNECT_TIMEOUT = 5000;
+        public static final long DEFAULT_NOOP_INTERVAL = 4000;
+        public static final int DEFAULT_BUFFER_SIZE = 4096;
+
+        private final String host;
+        private final String username;
+        private final String password;
+
+        private int port = DEFAULT_PORT;
+        private String serverCharset = DEFAULT_SERVER_CHARSET;
+        private int connectTimeout = DEFAULT_CONNECT_TIMEOUT;
+        private long noopInterval = DEFAULT_NOOP_INTERVAL;
+        private int bufferSize = DEFAULT_BUFFER_SIZE;
+
+        public Builder(String host, String username, String password) {
+            // 检查参数。
+            FtpConfigUtil.checkHost(host);
+            FtpConfigUtil.checkUsername(username);
+            FtpConfigUtil.checkPassword(password);
+
+            // 设置值。
+            this.host = host;
+            this.username = username;
+            this.password = password;
+        }
+
+        public Builder setPort(int port) {
+            FtpConfigUtil.checkPort(port);
+            this.port = port;
+            return this;
+        }
+
+        public Builder setServerCharset(String serverCharset) {
+            FtpConfigUtil.checkServerCharset(serverCharset);
+            this.serverCharset = serverCharset;
+            return this;
+        }
+
+        public Builder setConnectTimeout(int connectTimeout) {
+            FtpConfigUtil.checkConnectTimeout(connectTimeout);
+            this.connectTimeout = connectTimeout;
+            return this;
+        }
+
+        public Builder setNoopInterval(long noopInterval) {
+            FtpConfigUtil.checkNoopInterval(noopInterval, connectTimeout);
+            this.noopInterval = noopInterval;
+            return this;
+        }
+
+        public Builder setBufferSize(int bufferSize) {
+            FtpConfigUtil.checkBufferSize(bufferSize);
+            this.bufferSize = bufferSize;
+            return this;
+        }
+
+        @Override
+        public FtpConfig build() {
+            return new FtpConfig(
+                    host, port, username, password, serverCharset, connectTimeout, noopInterval, bufferSize
+            );
+        }
     }
 }